+++ CHANGES.md
@@ -7,6 +7,7 @@ Next Release (5.14.0)
 
 Features
 --------
+* [#1534](https://github.com/java-native-access/jna/pull/1534): Add `GetMethod`, `Put`, `SpawnInstance` to `c.s.j.p.win32.COM.WbemCli#IWbemClassObject` and `ExecMethod` to `c.s.j.p.win32.COM.WbemCli#IWbemServices` - [@faddom](https://github.com/faddom).
 
 Bug Fixes
 ---------